refactor: data 패키지 변동으로 인한 Dao test 패키지 변경
